UI: Implement drag and drop reordering for sources
Allows reordering of signals by dragging and dropping sources in the
list widget to the desired order.

Also adds missing QDataStream stream operators for the drag and drop; if
the data held within is not processed by the data stream, an assert will
occur, and QMetaType::save() will not work correctly.